[APIE-1098] Split the Flink CC/CP shared commands into isolated halves
`compute-pool` and `statement` are available under both a Confluent Cloud
and a Confluent Platform login, and each was a single constructor branching
on `cfg.IsCloudLogin()` with the two command sets, two output structs, and
the CMF-only helpers all in one file.

Split each into a cloud file and an on-prem file, and move the branch up to
`New()` in command.go. The branch has to exist somewhere: both halves define
`create`/`delete`/`describe`/`list`, and cobra resolves a child by name, so
only one set can ever be attached. `statement exception` is a shared group
one level down and is split the same way.

The parent Short now names its platform ("in Confluent Cloud" /
"in Confluent Platform") so each half owns its own string.

Also hoist the per-verb RunRequirement annotations to their parent command.
`ErrIfMissingRunRequirement` walks up from the leaf, so a single annotation
covers the subtree — for on-prem `statement` that is 9 annotations replaced
by 1, `exception list` included. This matches the placement already used by
`endpoint`, `artifact`, `connection` and `region`. The cloud `statement`
leaves carried no annotation and still carry none; that gap is left as-is.

Files moved, not rewritten: the CMF output structs and converters
(`computePoolOutOnPrem`, `statementOutOnPrem`, `exceptionOutOnPrem`,
`convertSdkComputePoolToLocalComputePool`,
`convertSdkStatementToLocalStatement`) now live beside the on-prem commands
that use them.
