+++ internal/flink/command_compute_pool.go
@@ -1,14 +1,9 @@
 package flink
 
 import (
-	"fmt"
-
 	"github.com/spf13/cobra"
 
-	cmfsdk "github.com/confluentinc/cmf-sdk-go/v1"
-
-	"github.com/confluentinc/cli/v4/pkg/config"
-	"github.com/confluentinc/cli/v4/pkg/flink"
+	pcmd "github.com/confluentinc/cli/v4/pkg/cmd"
 )
 
 type computePoolOut struct {
@@ -24,33 +19,20 @@ type computePoolOut struct {
 	Status      string `human:"Status" serialized:"status"`
 }
 
-type computePoolOutOnPrem struct {
-	CreationTime string `human:"Creation Time" serialized:"creation_time"`
-	Name         string `human:"Name" serialized:"name"`
-	Type         string `human:"Type" serialized:"type"`
-	Phase        string `human:"Phase" serialized:"phase"`
-}
-
-func (c *command) newComputePoolCommand(cfg *config.Config) *cobra.Command {
+func (c *command) newComputePoolCommand() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "compute-pool",
-		Short: "Manage Flink compute pools.",
+		Use:         "compute-pool",
+		Short:       "Manage Flink compute pools in Confluent Cloud.",
+		Annotations: map[string]string{pcmd.RunRequirement: pcmd.RequireNonAPIKeyCloudLogin},
 	}
 
-	if cfg.IsCloudLogin() {
-		cmd.AddCommand(c.newComputePoolCreateCommand())
-		cmd.AddCommand(c.newComputePoolDeleteCommand())
-		cmd.AddCommand(c.newComputePoolDescribeCommand())
-		cmd.AddCommand(c.newComputePoolListCommand())
-		cmd.AddCommand(c.newComputePoolUnsetCommand())
-		cmd.AddCommand(c.newComputePoolUpdateCommand())
-		cmd.AddCommand(c.newComputePoolUseCommand())
-	} else {
-		cmd.AddCommand(c.newComputePoolCreateCommandOnPrem())
-		cmd.AddCommand(c.newComputePoolDeleteCommandOnPrem())
-		cmd.AddCommand(c.newComputePoolDescribeCommandOnPrem())
-		cmd.AddCommand(c.newComputePoolListCommandOnPrem())
-	}
+	cmd.AddCommand(c.newComputePoolCreateCommand())
+	cmd.AddCommand(c.newComputePoolDeleteCommand())
+	cmd.AddCommand(c.newComputePoolDescribeCommand())
+	cmd.AddCommand(c.newComputePoolListCommand())
+	cmd.AddCommand(c.newComputePoolUnsetCommand())
+	cmd.AddCommand(c.newComputePoolUpdateCommand())
+	cmd.AddCommand(c.newComputePoolUseCommand())
 
 	return cmd
 }
@@ -61,35 +43,4 @@ func (c *command) validComputePoolArgs(cmd *cobra.Command, args []string) []stri
 	}
 
 	return c.autocompleteComputePools(cmd, args)
-}
-
-func convertSdkComputePoolToLocalComputePool(sdkComputePool cmfsdk.ComputePool) LocalComputePool {
-	localPool := LocalComputePool{
-		ApiVersion: sdkComputePool.ApiVersion,
-		Kind:       sdkComputePool.Kind,
-		Metadata: LocalComputePoolMetadata{
-			Name:              sdkComputePool.Metadata.Name,
-			CreationTimestamp: sdkComputePool.Metadata.CreationTimestamp,
-			Uid:               sdkComputePool.Metadata.Uid,
-			Labels:            sdkComputePool.Metadata.Labels,
-			Annotations:       sdkComputePool.Metadata.Annotations,
-		},
-		Spec: LocalComputePoolSpec{
-			Type:        sdkComputePool.Spec.Type,
-			ClusterSpec: sdkComputePool.Spec.ClusterSpec,
-		},
-	}
-
-	if phase := extractComputePoolPhase(sdkComputePool); phase != "" {
-		localPool.Status = &LocalComputePoolStatus{
-			Phase: phase,
-		}
-	}
-
-	return localPool
-}
-
-func extractComputePoolPhase(pool cmfsdk.ComputePool) string {
-	phase, _ := flink.GetMapField[string](pool.GetStatus(), "phase", fmt.Sprintf("compute pool %q", pool.GetMetadata().Name))
-	return phase
 }
