+++ internal/flink/command_compute_pool_create_onprem.go
@@ -18,11 +18,10 @@ import (
 
 func (c *command) newComputePoolCreateCommandOnPrem() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:         "create <resourceFilePath>",
-		Short:       "Create a Flink compute pool in Confluent Platform.",
-		Args:        cobra.ExactArgs(1),
-		Annotations: map[string]string{pcmd.RunRequirement: pcmd.RequireCloudLogout},
-		RunE:        c.computePoolCreateOnPrem,
+		Use:   "create <resourceFilePath>",
+		Short: "Create a Flink compute pool in Confluent Platform.",
+		Args:  cobra.ExactArgs(1),
+		RunE:  c.computePoolCreateOnPrem,
 	}
 
 	cmd.Flags().String("environment", "", "Name of the Flink environment.")
