+++ internal/flink/command_compute_pool_delete.go
@@ -19,7 +19,6 @@ func (c *command) newComputePoolDeleteCommand() *cobra.Command {
 		Short:             "Delete one or more Flink compute pools.",
 		Args:              cobra.MinimumNArgs(1),
 		ValidArgsFunction: pcmd.NewValidArgsFunction(c.validComputePoolArgsMultiple),
-		Annotations:       map[string]string{pcmd.RunRequirement: pcmd.RequireNonAPIKeyCloudLogin},
 		RunE:              c.computePoolDelete,
 	}
