+++ internal/flink/command_compute_pool_describe.go
@@ -14,7 +14,6 @@ func (c *command) newComputePoolDescribeCommand() *cobra.Command {
 		Short:             "Describe a Flink compute pool.",
 		Args:              cobra.MaximumNArgs(1),
 		ValidArgsFunction: pcmd.NewValidArgsFunction(c.validComputePoolArgs),
-		Annotations:       map[string]string{pcmd.RunRequirement: pcmd.RequireNonAPIKeyCloudLogin},
 		RunE:              c.computePoolDescribe,
 	}
