+++ internal/flink/command_compute_pool_describe_onprem.go
@@ -9,11 +9,10 @@ import (
 
 func (c *command) newComputePoolDescribeCommandOnPrem() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:         "describe <name>",
-		Short:       "Describe a Flink compute pool in Confluent Platform.",
-		Args:        cobra.ExactArgs(1),
-		Annotations: map[string]string{pcmd.RunRequirement: pcmd.RequireCloudLogout},
-		RunE:        c.computePoolDescribeOnPrem,
+		Use:   "describe <name>",
+		Short: "Describe a Flink compute pool in Confluent Platform.",
+		Args:  cobra.ExactArgs(1),
+		RunE:  c.computePoolDescribeOnPrem,
 	}
 
 	cmd.Flags().String("environment", "", "Name of the Flink environment.")
