+++ internal/flink/command_compute_pool_unset.go
@@ -12,12 +12,11 @@ import (
 
 func (c *command) newComputePoolUnsetCommand() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:         "unset",
-		Short:       "Unset the current Flink compute pool.",
-		Long:        "Unset the current Flink compute pool that was set with the `use` command.",
-		Args:        cobra.NoArgs,
-		RunE:        c.unset,
-		Annotations: map[string]string{pcmd.RunRequirement: pcmd.RequireNonAPIKeyCloudLogin},
+		Use:   "unset",
+		Short: "Unset the current Flink compute pool.",
+		Long:  "Unset the current Flink compute pool that was set with the `use` command.",
+		Args:  cobra.NoArgs,
+		RunE:  c.unset,
 		Example: examples.BuildExampleString(
 			examples.Example{
 				Text: `Unset default compute pool:`,
