+++ internal/flink/command_statement_describe_onprem.go
@@ -9,12 +9,11 @@ import (
 
 func (c *command) newStatementDescribeCommandOnPrem() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:         "describe [name]",
-		Short:       "Describe a Flink SQL statement.",
-		Long:        "Describe a Flink SQL statement in Confluent Platform.",
-		Args:        cobra.ExactArgs(1),
-		Annotations: map[string]string{pcmd.RunRequirement: pcmd.RequireCloudLogout},
-		RunE:        c.statementDescribeOnPrem,
+		Use:   "describe [name]",
+		Short: "Describe a Flink SQL statement.",
+		Long:  "Describe a Flink SQL statement in Confluent Platform.",
+		Args:  cobra.ExactArgs(1),
+		RunE:  c.statementDescribeOnPrem,
 	}
 
 	cmd.Flags().String("environment", "", "Name of the Flink environment.")
