+++ internal/flink/command_statement_exception.go
@@ -6,7 +6,6 @@ import (
 	"github.com/spf13/cobra"
 
 	pcmd "github.com/confluentinc/cli/v4/pkg/cmd"
-	"github.com/confluentinc/cli/v4/pkg/config"
 )
 
 type exceptionOut struct {
@@ -15,25 +14,16 @@ type exceptionOut struct {
 	Message   string    `human:"Message" serialized:"message"`
 }
 
-type exceptionOutOnPrem struct {
-	Timestamp string `human:"Timestamp" serialized:"timestamp"`
-	Name      string `human:"Name" serialized:"name"`
-	Message   string `human:"Message" serialized:"message"`
-}
-
-func (c *command) newStatementExceptionCommand(cfg *config.Config) *cobra.Command {
+func (c *command) newStatementExceptionCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "exception",
-		Short: "Manage Flink SQL statement exceptions.",
+		Short: "Manage Flink SQL statement exceptions in Confluent Cloud.",
 	}
 
-	if cfg.IsCloudLogin() {
-		pcmd.AddCloudFlag(cmd)
-		pcmd.AddRegionFlagFlink(cmd, c.AuthenticatedCLICommand)
-		cmd.AddCommand(c.newStatementExceptionListCommand())
-	} else {
-		cmd.AddCommand(c.newStatementExceptionListCommandOnPrem())
-	}
+	pcmd.AddCloudFlag(cmd)
+	pcmd.AddRegionFlagFlink(cmd, c.AuthenticatedCLICommand)
+
+	cmd.AddCommand(c.newStatementExceptionListCommand())
 
 	return cmd
 }
