+++ test/fixtures/output/flink/compute-pool/help.golden
@@ -1,4 +1,4 @@
-Manage Flink compute pools.
+Manage Flink compute pools in Confluent Cloud.
 
 Usage:
   confluent flink compute-pool [command]
