+++ test/fixtures/output/flink/statement/exception/help-onprem.golden
@@ -1,4 +1,4 @@
-Manage Flink SQL statement exceptions.
+Manage Flink SQL statement exceptions in Confluent Platform.
 
 Usage:
   confluent flink statement exception [command]
